requirements.txt
coverage==3.7.1
#https://www.djangoproject.com/download/1.7.b4/tarball/
django==1.6
tornado==4.0
sockjs-tornado==1.0
django-extensions==1.3.8
ipython==2.1
blinker==1.3
django-angular==0.7.3
djangorestframework==2.3.14
django-guardian==1.2.2
markdown==2.4.1
#-e git://github.com/bradjasper/django-jsonfield/tree/postgresql#egg=django-jsonfield
-e git://github.com/bradjasper/django-jsonfield#egg=django-jsonfield
celery==3.1.13
requests==2.3.0  # only for bots

